+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Replaced the `apply_status` function calls with `succ.apply_status` to handle status management in SUCC.
 * Combatant updates (marking combatant as defeated) also commented out as SUCC does this on its own.
 * Changed Incapacitation status to be an overlay (aka "big icon").
+* Actions: Added the `multiplyDmgMod` action to multiply the final damage.
 
 ## Version 2.75: Strange UI changes
 * Documentation: Brunocalado improved lots of documentation